requirements.txt
torch>=1.13
torchvision>=0.14
numpy>=1.21
Pillow>=9.0
tqdm>=4.64
matplotlib>=3.5
PyYAML>=6.0
